+++ httpie/cli/definition.py
@@ -501,7 +501,7 @@ output_options.add_argument(
     default=0,
     help='''
     Do not print to stdout or stderr, except for errors and warnings when provided once.
-    Provide twice to suppress errors and warnings.
+    Provide twice to suppress warnings as well.
     stdout is still redirected if --output is specified.
     Flag doesn't affect behaviour of download beyond not printing to terminal.